Add support for Latin glyphs
The exisitng Spyscape fonts do not include glyphs for characters with accents on then, as is needed for Spanish and other latin languages.

To overcome this issue, a new version of the Spyscape font has been produced, which includes new Latin glyphs.

This commit adds support for these new special characters.