Upgrade to Spyscape Two font, deprecate legacy Spyscape font
As of Feb 2019, a new typeface, "Spyscape Two" replaces the original "Spyscape" typeface. This includes redacted variants.

The new typefaces has been designed to improve legibility. It introduces new weights and variants. Because of this, it is
not a simple drop in replacement for the old font in some cases. Please see the brand guidelines on Frontify if in doubt
about useage.

This commit attempts to work as a drop in replacement where possible. Sass variables have therefore been remapped accordingly.
However, there may be cases where it is necessary to temporarily make use of the legacy typefaces. In these cases, the legacy
fonts have been moved to `$LEGACY_` variables. These will be removed in a future version once no longer used.